models/SBI/requirements.txt
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
albumentations==1.0.0
altair==5.0.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.1.0
Babel==2.14.0
backcall
backports.zoneinfo==0.2.1
bash-kernel==0.9.3
beautifulsoup4
bleach==6.0.0
blinker==1.6.3
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2021.5.30
cffi
chardet
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
comm==0.1.4
conda==4.10.1
conda-build==3.21.4
conda-package-handling
cryptography
cycler==0.11.0
debugpy==1.7.0
decorator
defusedxml==0.7.1
dlib==19.22.1
dnspython==2.1.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock
fonttools==4.38.0
fqdn==1.5.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
glob2
idna
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.17.0
Jinja2==3.1.2
joblib==1.3.2
jpeg4py==0.1.4
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-archive==3.4.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-events==0.6.3
jupyter-http-over-ws==0.0.8
jupyter-server==1.24.0
jupyter-server-fileid==0.9.1
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
libarchive-c
markdown-it-py==2.2.0
MarkupSafe
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.0
mkl-random
mkl-service==2.3.0
munch==4.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nbzip==0.1.0
nest-asyncio==1.5.8
networkx==2.6.3
notebook==6.5.6
notebook-shim==0.2.3
numpy
olefile==0.46
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==23.2
pandas==1.3.4
pandocfilters==1.5.0
parso
pexpect
pickleshare
Pillow==9.5.0
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
pluggy==1.2.0
pretrainedmodels==0.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil
ptyprocess
pyarrow==12.0.1
pycosat==0.6.3
pycparser
pydeck==0.8.1b0
pygments==2.17.2
Pympler==1.0.1
pyOpenSSL
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks
pytest==7.4.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pytz
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
requests==2.31.0
retinaface-pytorch==0.0.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
ruamel-yaml-conda
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.1
Send2Trash==1.8.2
shapely==2.0.2
six
smmap==5.0.1
sniffio==1.3.0
soupsieve
streamlit==1.23.1
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.9.0+cu111
torchelastic==0.2.0
torchtext==0.10.0
torchvision==0.10.0+cu111
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3
validators==0.20.0
watchdog==3.0.0
wcwidth
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.9
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.15.0
